allowSnapshots from app version Backup (#1215)
* allowSnapshots from app version Backup

Existence of a Backup in an app version enables snapshots unless
overridden by license or by kots.io/exclude or kots.io/when annotations
in the Backup itself.

* Use the deployed version's Backup

* lookup downstream by cluster ID

There may be multiple downstreams with different current sequences. For
the allowSnapshots function use the first downstream, matching the logic
for the version in the dashboard card on the app detail page.