Download snapshot logs (#1242)
* Download snapshot logs

The handler is implemented in the Go API server and is authenticated
with a web session Auth header. It creates a velero/v1/DownloadRequest
and then watches it until the velero operator has added a signed URL to
its status where the logs can be downloaded. The handler then deletes
the DownloadRequest and pipes the logs file from the backend object
store to the browser.

The client currently uses fetch and does not treat the response as a
file. That will need to be fixed in another PR.