Support non-access key based AWS credentials (#2032)
* Support non-access key based AWS credentials

Currently kots determines if the Docker repo being used is an ECR repo. If it is, it expects to be provided with an access key and secret to authenticate with the ECR repo. However in some cases, kots will be run from an assumed role or EC2 instance that doesn't have an access key. The go SDK will use a default credential provider chain if no credentials are specified that will use the instance role or other credentials.

This change allows kots to be installed without an access key when using ECR

* Update pkg/docker/registry/ecr.go

Co-authored-by: Salah Aldeen Al Saleh <sg.alsaleh@gmail.com>

* Code review feedback

* Update cmd/kots/cli/admin-console-push-images.go

Co-authored-by: Salah Aldeen Al Saleh <sg.alsaleh@gmail.com>

* remove unneeded parens

Co-authored-by: Salah Aldeen Al Saleh <sg.alsaleh@gmail.com>