Generate Signed SBOM (#2121)
This PR contains build related changes that will generate an SBOM (Software Bill of Materials) for the Javascript dependencies of the project, another for Go dependencies, and combines them into a tar file. The tar is signed and both the SBOM tar file, and it's signature are included in releases.