bug: setting https proxy to the right value
in order to handle two distinct proxies that may exist in the system
separately, it is necessary to treat them individually. the presence
of a second environment variable, PROXY_HTTPS_ADDRESS, indicates the
configured https proxy.

this commit ensures the appropriate utilization of the correct env
variable for the https proxy.