bug: do not fail to install if no embedded cluster installation
if no embedded cluster installation object has been found in the cluster
we can't start an upgrade. on this case (where the installation does not
exists) we shouldn't be failing but moving on instead.

during the initial cluster bootstrap the installation object may take a
while to be created therefore a scenario where the kots app is installed
but no installation object exists is valid.