bug: stop fetching support bundle uri on airgap
the function `ParseSupportBundleFromDoc()` always attempt to fetch the
support bundle from its uri (if present). there is another function that
allows us to define if we want or not use the uri:

```
ParseSupportBundle()
```

this new function requires a second argument called `followURI` and we
need to set it to false when kots is running in an airgap environment.