Adds missing fields to `LicenseFieldValue` (#5698)
TL;DR
-----

Updates the `LicenseFieldValues` template function for fields that
weren't included

Details
-------

Enables `LicenseFieldValue` for fields it wasn't able to provide values
for before. Also rearranges the ordering to match how they are grouped
in the docs.

N.B. There are fields in the docs that `kotskinds` doesn't track and
that I didn't see in a sample license, so I've left those out.